+++ test/PHPUnit/class/actes/ActesTransactionsCloserTest.php
@@ -1,5 +1,6 @@
 <?php
 
+use PHPUnit\ActesUtilitiesTestTrait;
 use S2lowLegacy\Class\actes\ActesStatusSQL;
 use S2lowLegacy\Class\actes\ActesTransactionsCloser;
 use S2lowLegacy\Class\actes\ActesTransactionsSQL;